TBA: Multiplication Table Format modification (this commit modifies only the 5d, 2d and 3d will follow)
git-svn-id: svn+ssh://erebor.esa.informatik.tu-darmstadt.de/home/wimi/svn/GaalopCompiler/gaalop-2.0/branches/tbaAndGapp@908 4235ee25-d3ef-403b-b6ed-692c26853d4b

+++ tba/src/main/java/de/gaalop/tba/cfgImport/MvExpressionsBuilder.java
@@ -141,7 +141,7 @@ public class MvExpressionsBuilder extends EmptyControlFlowVisitor implements Exp
 						Expression prodExpr = new Multiplication(left.bladeExpressions[bladeL],right.bladeExpressions[bladeR]);
 						Multivector prodMv = usedAlgebra.getProduct(typeProduct,bladeL,bladeR);
 
-						double[] prod = prodMv.getValueArr();
+						byte[] prod = prodMv.getValueArr();
 
 						for (int bladeResult = 0;bladeResult<bladeCount;bladeResult++)
 							if (Math.abs(prod[bladeResult])>EPSILON) {
